remove base views
