--- lib/locales/en/commands/base.yml
@@ -1,6 +1,0 @@
-en:
-  commands:
-    base:
-      missing_positional_argument: "Missing #%{index} positional argument: '%{name}'"
-      invalid_subcommand: "Invalid subcommand: '%{name}'"
-      subcommand_needs_positionals: "This subcommand should be called with positionals"
