+++ lib/locales/pt-BR/argument.yml
@@ -1,0 +1,6 @@
+pt-BR:
+  argument:
+    missing_positional_argument: "Faltando o argumento posicional #%{index}: '%{name}'"
+    invalid_subcommand: "Subcomando inválido: '%{name}'"
+    subcommand_needs_positionals: "Este subcomando deve ser chamado com argumentos posicionais"
+    invalid_category: "Categoria inválida '%{name}'. Use 'l' ou 's'"
